refactor(mobile): retire the single Board look screen
The three screens are live, so the one they replace and the sections only it
rendered are dead. Accessibility becomes the second leaf: the same controls, one
level down, wrapping the existing section rather than rewriting it — the split is
about where these live, not what they do, and the colour picker underneath is not
worth disturbing to make a point about structure.

Coverage moves rather than going away. The old suite needed nearly three hundred
lines of mocks to assert which rows a screen renders; most of that is now a pure
model test with no renderer at all. What genuinely needed a rendered tree — the
marker shape controls appearing only when the drawing is KNOWN to be Classic,
including the case where the capability probe has not answered yet — stays as a
rendered test against the leaf that now owns those rows.

The two bugs the shared writer fixed get tests of their own, because neither is
visible in a screenshot: role glyphs reaching the remembered look, a restore
being unable to switch them off, and a board-look reset leaving hold colours
alone.
