+++ packages/mobile/app/(tabs)/profile/board-look/accessibility.tsx
@@ -1,0 +1,5 @@
+import { BoardLookAccessibilityScreen } from '../../../../src/components/settings/board-look/BoardLookAccessibilityScreen';
+
+export default function BoardLookAccessibilityRoute() {
+  return <BoardLookAccessibilityScreen />;
+}
